refactor(mediawiki): split quality.go into terminology/translation/audit siblings (deslop #59)
